+++ corewa_rs/src/cli.rs
@@ -8,7 +8,7 @@ use std::{
 use lazy_static::lazy_static;
 use structopt::StructOpt;
 
-use crate::phased_parser::parse;
+use crate::parser::parse;
 
 lazy_static! {
     static ref IO_SENTINEL: PathBuf = PathBuf::from("-");
